Moving rest services into AuthenticationControllerBrapi
ISSUE: IBP-3188
