+++ src/main/java/org/ibp/api/brapi/v1/security/auth/AuthenticationControllerBrapi.java
@@ -7,6 +7,8 @@ import org.ibp.api.brapi.v1.common.Metadata;
 import org.ibp.api.security.xauth.Token;
 import org.ibp.api.security.xauth.TokenProvider;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
@@ -17,9 +19,13 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import java.io.UnsupportedEncodingException;
+import java.net.URI;
 import java.net.URL;
+import java.net.URLEncoder;
 
 /**
  * BMS implementation of the <a href="http://docs.brapi.apiary.io/">BrAPI</a> Authentication services.
@@ -53,6 +59,16 @@ public class AuthenticationControllerBrapi {
 
 		return new TokenResponse(new Metadata(null, null, new URL[] {}), username, token.getToken(), token.getExpires());
 	}
+
+	@RequestMapping(value = "/brapi/authorize", method = RequestMethod.GET)
+	@ResponseBody
+	public ResponseEntity authorize(@RequestParam(value = "display_name") final String display_name,@RequestParam(value = "return_url") final String return_url)
+		throws UnsupportedEncodingException {
+		final URI loginUrl = URI.create(
+			"/ibpworkbench/controller/auth/login?display_name=" + URLEncoder.encode(display_name,"UTF-8") + "&return_url=" + return_url);
+		return ResponseEntity.status(HttpStatus.MOVED_PERMANENTLY).location(loginUrl).build();
+	}
+
 
 	public void setTokenProvider(final TokenProvider tokenProvider) {
 		this.tokenProvider = tokenProvider;
